docker container works just fine

+++ migrations/20260307180845_create_product_prices_table.sql
@@ -12,7 +12,7 @@ CREATE TABLE product_prices (
     currency VARCHAR(10) NOT NULL,
     amount_in_base_currency REAL,
 
-    original_amount REAL,
+    original_amount BIGINT,
     discount_percentage REAL,
 
     created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
